moved toggle lsm
